log parser function `log2var` works
The log2var function is defined and works when called on another
projects. Right now the way to use the result is through a callback
since it's node and it's asynchronous.

The asociated jest test file works with it's imports, it did not work
before. The test are WIP.

eslint has also been added and everything still works.